Attendace_Backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database & Supabase (supabase 2.9+ with httpx 0.26+ fixes gotrue 'proxy' TypeError)
supabase>=2.9.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Config
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# ML & Face Recognition (InsightFace: RetinaFace + ArcFace, state-of-the-art)
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python-headless>=4.9.0
mediapipe>=0.10.9
numpy>=1.26.0,<2.0
Pillow>=10.0.0

# SMS & Async
httpx>=0.26
apscheduler==3.10.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Utilities
structlog==24.1.0
slowapi==0.1.9
python-json-logger==2.0.7
